Website Blocker
This is an GUI application that blocks certain websites that are harmful for our PC and helps to save our time. We have build this application in python with the help of TKinter library. An small application designed to make people show that how to save time and be ware  from fraudulents calls from the internet.